+++ harmony-server/rest/v1/addchannel.go
@@ -1,0 +1,11 @@
+package v1
+
+import (
+	"github.com/labstack/echo/v4"
+	"golang.org/x/time/rate"
+)
+
+func AddChannel(limiter *rate.Limiter, ctx echo.Context) error {
+
+	return nil
+}
