+++ harmony-server/rest/v1/fileupload.go
@@ -1,4 +1,4 @@
-package rest
+package v1
 
 import (
 	"fmt"
