+++ harmony-server/rest/v1/getchannels.go
@@ -1,4 +1,4 @@
-package rest
+package v1
 
 import (
 	"github.com/labstack/echo/v4"
